fix: Improve disconnect logic for ssh-bin
Previously we were only closing stdin, and normally that's enough, but
not always: if the connection is gone, but ssh doesn't know about it
(e.g. if the OS was suspended for long enough time, and then resumed),
then the connection kept hanging.

This commit provides custom context for the external ssh command, and
cancels it when we need to disconnect.